backend/requirements.txt
fastapi==0.122.0
uvicorn==0.38.0
python-multipart==0.0.20
pydantic==2.12.4
pdfplumber==0.11.8
requests==2.32.5
python-dotenv==1.2.1
openai==2.8.1
langchain==1.1.0
langchain-groq==1.1.0
pytest==9.0.1
pytest-asyncio==1.3.0
